Update external neopixel blocks
