+++ libs/light/create.ts
@@ -1,0 +1,32 @@
+
+namespace light {
+    /**
+     * Create a new NeoPixel driver for `numleds` LEDs.
+     * @param pin the pin where the neopixel is connected.
+     * @param numleds number of leds in the strip, eg: 24,30,60,64
+     */
+    //% blockId="neopixel_create" block="create strip|pin %pin|pixels %numleds|mode %mode"
+    //% help="light/create-neo-pixel-strip"
+    //% trackArgs=0,2
+    //% parts="neopixel"
+    //% weight=100
+    export function createNeoPixelStrip(
+        pin: DigitalPin = null,
+        numleds: number = 10,
+        mode?: NeoPixelMode
+    ): NeoPixelStrip {
+        if (!mode)
+            mode = NeoPixelMode.RGB
+
+        const strip = new NeoPixelStrip();
+        strip._mode = mode;
+        strip._length = Math.max(0, numleds);
+        strip._start = 0;
+        strip._pin = pin ? pin : defaultPin();
+        strip._pin.digitalWrite(false);
+        strip._barGraphHigh = 0;
+        strip._barGraphHighLast = 0;
+        strip.setBrightness(20)
+        return strip;
+    }
+}
